backend/requirements.txt
apturl==0.5.2
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
axios==0.4.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
channels==4.1.0
chardet==4.0.0
charset-normalizer==2.0.6
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
constantly==23.10.4
cryptography==42.0.7
cupshelpers==1.0
daphne==4.1.2
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
django-cors-headers==4.3.1
django-filter==2.4.0
django-shortuuidfield==0.1.3
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-extensions==0.7.1
duplicity==0.8.21
exceptiongroup==1.2.0
fasteners==0.14.1
filelock==3.13.4
future==0.18.2
gyp==0.1
httpie==2.6.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==22.10.0
iniconfig==2.0.0
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.2.1
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==24.0
paramiko==2.9.3
pbr==6.0.0
pexpect==4.8.0
Pillow==8.3.2
pip==22.0.2
platformdirs==4.2.0
pluggy==1.4.0
protobuf==3.12.4
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
pytest==8.1.1
pytest-django==4.4.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
rich==13.7.1
SecretStorage==3.3.1
service-identity==24.1.0
setuptools==59.6.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
stevedore==5.2.0
systemd-python==234
tomli==2.0.1
ttystatus==0.38
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.25.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
virtualenvwrapper-win==1.2.7
wadllib==1.3.6
wheel==0.37.1
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==6.4.post2
